\name{get_tweets}
\alias{get_tweets}
\title{Grab tweets for a particular user}
\usage{
  get_tweets(user, consumer.key, consumer.secret,
    access.token, access.token.secret, n.tweets = 300)
}
\arguments{
  \item{user}{the name of a Twitter user}

  \item{consumer.key}{your Twitter "consumer" key, see
  Details.}

  \item{consumer.secret}{your Twitter "consumer" secret,
  see Details.}

  \item{access.token}{access token}

  \item{access.token.secret}{access token secret}

  \item{n.tweets}{number of tweets to pull from the user}
}
\description{
  Here is where some blurb goes.
}
\section{Details

First need to set up a "new app" get key and secret...
https}{
  //dev.twitter.com/apps/new
}
\author{
  David L. Miller
}

