\name{neigh}
\alias{neigh}
\title{Generate a string from a horse object}
\usage{
  neigh(horse.obj, length = 140)
}
\arguments{
  \item{horse.obj}{a horse object containing the transition
  matrix for the Markov chain.}

  \item{length}{maximum length for the resulting string
  (default 140).}
}
\value{
  a string of length \code{length}.
}
\description{
  This function generates strings from horse objects. Once
  the transition matrix has been found for your corpus
  (e.g. using \code{\link{get_tweets}} you can use the
  resulting horse object (generated by
  \code{\link{make_horse}}) to generate new strings.
}
\author{
  David L. Miller
}

