\name{twitter_setup}
\alias{twitter_setup}
\title{Setup a connection to the Twitter API}
\usage{
  twitter_setup(consumer.key, consumer.secret,
    access.token, access.token.secret)
}
\arguments{
  \item{consumer.key}{your Twitter "consumer" key, see
  Details.}

  \item{consumer.secret}{your Twitter "consumer" secret,
  see Details.}

  \item{access.token}{access token}

  \item{access.token.secret}{access token secret}
}
\description{
  Here is where some blurb goes.
}
\section{Details

First need to set up a "new app" get key and secret...
https}{
  //dev.twitter.com/apps/new
}
\author{
  David L. Miller
}

